🧹 Tidy: useData.tsのリファクタリングと型安全化
