+++ .jules/tidy.md
@@ -76,3 +76,14 @@
 2024-05-24 - [カスタムフックによる非同期アクションの共通化]
 学び: ダッシュボードの複数のウィジェット(Feeding, Diaper, Sleep)で、API呼び出し時のローディング状態管理、エラーハンドリング、トースト通知のロジックが重複していた。これらはUIの違いはあるものの、「非同期アクションを実行し、結果をユーザーに通知する」という振る舞いは同一だった。
 アクション: `useAsyncAction` フックを作成し、これらの共通ロジックを抽出した。これにより、コンポーネント側は「何をするか（API呼び出し）」と「どう通知するか（メッセージ）」に集中でき、状態管理のボイラープレートを排除できた。今後も同様のパターン（設定変更画面など）でこのフックを再利用できる可能性がある。
+
+2026-02-23 - [型定義] データフェッチ用フックの型安全化と共通化
+
+学び:
+- `useData.ts` 内でインラインの `as` キャストが多用されており、型安全性が損なわれていた。特にAPIの変更に対する耐性が低かった。
+- OpenAPIから生成された型定義 (`FamilyResponse` 等) があるにもかかわらず、手動で定義された部分的なインターフェース (`Member` 等) がコンポーネント内に散在しており、整合性が取れていなかった。
+- `useBabyResource` のような共通ロジックを独立したフックとして切り出すことで、再利用性とテスト容易性が向上した。
+
+アクション:
+- 今後新しいデータフェッチフックを作成する際は、必ず `types/generated/api.d.ts` の型を使用し、インラインでの型定義を避ける。
+- コンポーネント内でローカルに定義されているインターフェースを見つけたら、共通の型定義に置き換えるリファクタリングを継続する。
