+++ frontend/components/dashboard/GrowthWidget.tsx
@@ -1,7 +1,7 @@
 "use client"
 import { memo } from "react"
 import { areRecordsEqual } from "@/lib/memoUtils"
-import { BabyRecord } from "@/hooks/useData"
+import { BabyRecord } from "@/types/record"
 import { WidgetCard } from "./WidgetCard"
 import { BabyBottleLoading } from "@/components/ui/baby-bottle-loading"
