+++ frontend/components/dashboard/QuickActionBar.tsx
@@ -3,7 +3,7 @@ import { useState, useMemo } from "react"
 import { Button } from "@/components/ui/button"
 import { api } from "@/lib/api"
 import { usePermissions } from "@/hooks/usePermissions"
-import { BabyRecord } from "@/hooks/useData"
+import { BabyRecord } from "@/types/record"
 import { Moon, Bed, Loader2, StickyNote } from "lucide-react"
 import { useRecordFeedback } from "@/hooks/useRecordFeedback"
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog"
