+++ frontend/types/family.ts
@@ -1,0 +1,4 @@
+import type { components } from "@/types/generated/api"
+
+export type Family = components["schemas"]["FamilyResponse"]
+export type FamilyMember = components["schemas"]["FamilyMemberResponse"]
